requirements.txt
# Discord bot core
discord.py==2.4.0

# HTTP / сети
aiohttp==3.11.8
requests==2.32.3
httpx==0.27.2
websockets==12.0
discord-webhook==1.3.1

# Конфиги
PyYAML==6.0.2
bestconfig==1.2.0
python-dotenv==1.0.1

# SSH
paramiko==3.5.0
cryptography>=42.0.5

# OpenAI
openai>=1.43.0,<2.0.0

# Работа с БД (SQLAlchemy)
SQLAlchemy==2.0.23
pymysql==1.1.0

# Изображения (используется в yandexgptart)
Pillow==10.4.0

# Прочее
loguru==0.7.2
